# GestorBarberia.Api/bin/Release/net6.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "GestorBarberia.Api, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/Administrador/GetAdministradorBarberia/{name}/{password}": {
      "get": {
        "tags": [
          "Administrador"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Administrador/GetAdministradorApp/{name}/{password}": {
      "get": {
        "tags": [
          "Administrador"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Administrador/GetAdministradorById/{id}": {
      "get": {
        "tags": [
          "Administrador"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Administrador/Save": {
      "post": {
        "tags": [
          "Administrador"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AdminAddDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Administrador/Update": {
      "put": {
        "tags": [
          "Administrador"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AdminUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Administrador/Delete": {
      "delete": {
        "tags": [
          "Administrador"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/AdminRemoveDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Barberia/Save": {
      "post": {
        "tags": [
          "Barberia"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BarberiaAddDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Barberia/Update": {
      "put": {
        "tags": [
          "Barberia"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BarberiaUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Barberia/Delete": {
      "delete": {
        "tags": [
          "Barberia"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BarberiaRemoveDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Barbero/GetBarbero/{name}/{password}": {
      "get": {
        "tags": [
          "Barbero"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "name",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "password",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Barbero/BarberosByBarberiaId/{id}": {
      "get": {
        "tags": [
          "Barbero"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Barbero/BarberosByClienteBarberiaId/{id}": {
      "get": {
        "tags": [
          "Barbero"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Barbero/BarberoById/{id}": {
      "get": {
        "tags": [
          "Barbero"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Barbero/Save": {
      "post": {
        "tags": [
          "Barbero"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BarberoAddDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Barbero/Update": {
      "put": {
        "tags": [
          "Barbero"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BarberoUpdateDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Barbero/Delete": {
      "delete": {
        "tags": [
          "Barbero"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/BarberoRemoveDto"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Barbero/DeleteByAdmin/{barberoId}/{adminId}": {
      "delete": {
        "tags": [
          "Barbero"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "barberoId",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "path",
            "name": "adminId",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "definitions": {
    "AdminAddDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "administradoresId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "tipo": {
          "type": "string"
        }
      }
    },
    "AdminRemoveDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "administradoresId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "tipo": {
          "type": "string"
        }
      }
    },
    "AdminUpdateDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "administradoresId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "tipo": {
          "type": "string"
        }
      }
    },
    "BarberiaAddDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "barberiasId": {
          "format": "int32",
          "type": "integer"
        },
        "nombreBarberia": {
          "type": "string"
        },
        "admin": {
          "format": "int32",
          "type": "integer"
        },
        "nombreAdministrador": {
          "type": "string"
        }
      }
    },
    "BarberiaRemoveDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "barberiasId": {
          "format": "int32",
          "type": "integer"
        },
        "nombreBarberia": {
          "type": "string"
        },
        "admin": {
          "format": "int32",
          "type": "integer"
        },
        "nombreAdministrador": {
          "type": "string"
        }
      }
    },
    "BarberiaUpdateDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "barberiasId": {
          "format": "int32",
          "type": "integer"
        },
        "nombreBarberia": {
          "type": "string"
        },
        "admin": {
          "format": "int32",
          "type": "integer"
        },
        "nombreAdministrador": {
          "type": "string"
        }
      }
    },
    "BarberoAddDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "barberoId": {
          "format": "int32",
          "type": "integer"
        },
        "barberiaId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "imgbarbero": {
          "type": "string"
        }
      }
    },
    "BarberoRemoveDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "barberoId": {
          "format": "int32",
          "type": "integer"
        },
        "barberiaId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "imgbarbero": {
          "type": "string"
        }
      }
    },
    "BarberoUpdateDto": {
      "type": "object",
      "properties": {
        "changeDate": {
          "format": "date-time",
          "type": "string"
        },
        "changeUser": {
          "format": "int32",
          "type": "integer"
        },
        "barberoId": {
          "format": "int32",
          "type": "integer"
        },
        "barberiaId": {
          "format": "int32",
          "type": "integer"
        },
        "nombre": {
          "type": "string"
        },
        "telefono": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "imgbarbero": {
          "type": "string"
        }
      }
    },
...
  }
}