# GestorBarberia.Application/Dtos/BarberoDto/BarberoUpdateDto.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace GestorBarberia.Application.Dtos.BarberoDto
{
    public class BarberoUpdateDto : BarberoDto
    {
    }
}